chore: move RelatedConfig to tests from models
